Revert the “styled” naming convention thing
It’s too many files. It’s confusing. I take advantage of hoisting; make
sure to put styled stuff at the end.
