--- app/src/pages/AbstractStreamEditPage/PartitionsSection/Partitions.jsx
@@ -1,7 +1,0 @@
-import styled from 'styled-components'
-
-const Partitions = styled.div`
-    max-width: 136px;
-`
-
-export default Partitions
