--- app/src/pages/AbstractStreamEditPage/PartitionsSection/index.jsx
@@ -1,11 +1,0 @@
-import styled from 'styled-components'
-import UnstyledPartitionsSection from './index.unstyled'
-
-const PartitionsSection = styled(UnstyledPartitionsSection)`
-    > p {
-        margin-bottom: 3.125rem;
-        max-width: 660px;
-    }
-`
-
-export default PartitionsSection
