--- app/src/shared/components/Ui/Label.unstyled.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function Label({ className, children }) {
-    return (
-        <label className={className}>
-            {children}&zwnj;
-        </label>
-    )
-}
